Chore: refactored store of Trips
